Split generic GUI tests into focused modules
